_posts/fsgai/hello_world/gemini_hello.py
from google import genai

client = genai.Client(
    api_key="AIzaSyBjA34ENgeGNplvIqCP-qcH2fuMkqxdO7o"
)

response = client.models.generate_content(
    model="gemini-2.5-flash", contents="Explain how AI works in a few words"
)

print(response.text)